Make small edits to grammar, consistency and style
This proposed set of edits fixes a few minor items (a comma splice here, a missing comma there), removes a few explanation points (the original text had quite a few), makes headings more consistent, and adds some clarifying text here and there. It also flips the order of the two topics listed in the very first sentence because this new order seems more logical to me, but if you want to keep the original order, that would be fine with me too.