Replace {hash} with X in explanations
This is maybe controversial, but I find that adding any kind of
special characters like { } in explanations confuses people, and they
have a tendency to think that they should type those special
characters.  A solution I wish we could use is to use special
formatting for the value of X in the text `ipfs get X`, to make it
really clear it's mean as a variable.  Failing that, I couldn't think
of anything clearer or less confusing that just using "X" in this
case.

People may have experiences doing this in a different/better way, in
which case, feel free to ignore this change.